requirements.txt
text-generation
langchain
transformers
runpod
python-dotenv
gradio
scipy
elevenlabs
openai
torch
wandb
word2number